refactor(dashboard): extract Mastio agents sub-router (F-B-201 PR-4/10)
Audit ref: imp/audits/2026-05-20/findings/track-b/F-B-201.md

Sprint F-B-201 PR-4 of 10. base=main. Mirrors Court PR-7 (#850)
pattern on the Mastio dashboard, kept in one sub-router because the
Mastio agent surface is simpler than Court's (no separate lifecycle
vs credentials split).

Extracts 7 routes (agent registry list + per-agent management) plus
one private helper from mcp_proxy/dashboard/router.py.

  mcp_proxy/dashboard/agents_routes.py (new, 452 LOC, 7 routes):
    GET  /proxy/agents                        list + KPIs + org-status badge
    POST /proxy/agents/create                 enroll a new internal agent
    GET  /proxy/agents/{agent_id}             developer-portal detail page
    GET  /proxy/agents/{agent_id}/env-download download env.sample
    POST /proxy/agents/{agent_id}/reach       edit reach (intra/cross/both)
    POST /proxy/agents/{agent_id}/deactivate  flip is_active=False
    POST /proxy/agents/{agent_id}/delete      permanent delete with approval-hook
    + private _refresh_org_status_from_broker (was inline at router.py:324,
      only consumer was agents_page)

  mcp_proxy/dashboard/router.py (4117 -> 3720 LOC, -397 net):
    - 7 route bodies + 1 private helper removed; breadcrumb at the old
      positions.
    - One new include_router(_agents_routes.router) call.

Behavioural delta: zero. ACTION_AGENTS_DELETE approval-hook gate
preserved on POST /agents/{id}/delete. _VALID_REACH whitelist
({intra,cross,both}) preserved on POST /agents/{id}/reach.
agents_page still polls broker on hot reload via
_refresh_org_status_from_broker.

Tests:
  - pytest tests/test_dashboard_agents_reach.py
    tests/test_dashboard_approvals_nav.py
    tests/test_dashboard.py
    tests/test_dashboard_principals.py
    tests/test_local_token_user_session.py
    tests/test_wave_b_pr7_local_token_dpop_binding.py
    --deselect tests/test_dashboard_approvals_nav.py::test_badge_approvals_empty_when_plugin_unavailable
    -n auto --dist=loadfile -q
  - 83 passed (1 pre-existing env-fail deselected: enterprise package
    must not be installed in tree; passes clean in CI).

Router LOC trajectory:
  Pre-Sprint:    66 routes, 5106 LOC
  Post-PR-1 #859: 66 routes, 5052 LOC
  Post-PR-2 #865: 61 routes, 4828 LOC
  Post-PR-3 #866: 58 routes, 4117 LOC
  Post-PR-4 (this): 51 routes, 3720 LOC
  Target post-PR-10: ~3 routes, ~120 LOC aggregator.

Next PR-5 (F-B-201 modularization): /proxy/tools + /proxy/network
+ /proxy/tools/reload (3 routes, ~140 LOC) into tools_network_routes.py.
